+++ ScalablePtrView/src/com/aj2014/scalableptrview/IPtrView.java
@@ -12,7 +12,7 @@ public interface IPtrView {
 	 * 
 	 */
 	public void switchTo(EPtrState state);
-	public void onPull(int distance);
+	public void onPull(float distance);
 	public boolean recover(int distance);
 	
 }
