+++ ScalablePtrView/src/com/aj2014/scalableptrview/IScalableView.java
@@ -5,7 +5,7 @@ public interface IScalableView {
 	/**
 	 * 缩放到指定值
 	 */
-	public void scaleTo(int distance);
+	public void scaleTo(float distance);
 	/**
 	 * 状态恢复
 	 * @param distance scale 总距离
